+++ hive/test/integration/put_large_strings_test.dart
@@ -2,16 +2,16 @@ import 'package:test/test.dart';
 
 import 'integration.dart';
 
-Future _performTest(bool lazy) async {
-  var box = await openBox(lazy);
+Future _performTest(bool lazy, {required bool isolated}) async {
+  var (hive, box) = await openBox(lazy, isolated: isolated);
   for (var i = 0; i < 5; i++) {
     final largeString = i.toString() * 1000000;
     await box.put('string$i', largeString);
   }
 
-  box = await box.reopen();
+  box = await hive.reopenBox(box);
   for (var i = 0; i < 5; i++) {
-    final largeString = await await box.get('string$i');
+    final largeString = await box.get('string$i');
 
     expect(largeString == i.toString() * 1000000, true);
   }
@@ -19,13 +19,15 @@ Future _performTest(bool lazy) async {
 }
 
 void main() {
-  group(
-    'put large strings',
-    () {
-      test('normal box', () => _performTest(false));
+  hiveIntegrationTest((isolated) {
+    group(
+      'put large strings',
+      () {
+        test('normal box', () => _performTest(false, isolated: isolated));
 
-      test('lazy box', () => _performTest(true));
-    },
-    timeout: longTimeout,
-  );
+        test('lazy box', () => _performTest(true, isolated: isolated));
+      },
+      timeout: longTimeout,
+    );
+  });
 }