+++ hive/test/integration/put_many_strings_test.dart
@@ -4,33 +4,35 @@ import '../tests/frames.dart';
 import '../util/is_browser.dart';
 import 'integration.dart';
 
-Future _performTest(bool lazy) async {
-  final hive = await createHive();
+Future _performTest(bool lazy, {required bool isolated}) async {
+  final hive = await createHive(isolated: isolated);
   final repeat = isBrowser ? 20 : 1000;
-  var box = await openBox(lazy, hive: hive);
+  var (_, box) = await openBox(lazy, isolated: isolated, hive: hive);
   for (var i = 0; i < repeat; i++) {
     for (final frame in valueTestFrames) {
       await box.put('${frame.key}n$i', frame.value);
     }
   }
 
-  box = await box.reopen();
+  box = await hive.reopenBox(box);
   for (var i = 0; i < repeat; i++) {
     for (final frame in valueTestFrames) {
-      expect(await await box.get('${frame.key}n$i'), frame.value);
+      expect(await box.get('${frame.key}n$i'), frame.value);
     }
   }
   await box.close();
 }
 
 void main() {
-  group(
-    'put many strings',
-    () {
-      test('normal box', () => _performTest(false));
+  hiveIntegrationTest((isolated) {
+    group(
+      'put many strings',
+      () {
+        test('normal box', () => _performTest(false, isolated: isolated));
 
-      test('lazy box', () => _performTest(true));
-    },
-    timeout: longTimeout,
-  );
+        test('lazy box', () => _performTest(true, isolated: isolated));
+      },
+      timeout: longTimeout,
+    );
+  });
 }